Match sub_802E164 and sub_802E1EC

+++ include/game/game.h
@@ -606,7 +606,7 @@ void Player_SetMovestate_IsInScriptedSequence(void);
 void Player_ClearMovestate_IsInScriptedSequence(void);
 
 void sub_802EFDC(u32);
-void sub_802E164(u16, u16);
+void sub_802E164(s32, u16);
 
 // ApplyGamePlaySettings
 void ApplyGameStageSettings(void);
